# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Las Vegas Reservation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
        - name: roomId
          in: query
          required: false
          schema:
            type: integer
        - name: checkInDate
          in: query
          required: false
          schema:
            type: string
        - name: checkOutDate
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of reservations with full details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservationDetail"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "400":
          description: Bad request or room not available
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "404":
          description: Not found
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Delete a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /reservations/{id}/checkin:
    patch:
      operationId: checkInReservation
      tags: [reservations]
      summary: Check in a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "404":
          description: Not found
  /reservations/{id}/checkout:
    patch:
      operationId: checkOutReservation
      tags: [reservations]
      summary: Check out a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "404":
          description: Not found
  /reservations/{id}/cancel:
    patch:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationDetail"
        "404":
          description: Not found
  /reservations/export/csv:
    get:
      operationId: exportReservationsCSV
      tags: [reservations]
      summary: Export all reservations as CSV for Google Sheets
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, number, type, floor, status, pricePerNight, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        type:
          type: string
        floor:
          type: integer
        status:
          type: string
        pricePerNight:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Employee:
      type: object
      required: [id, name, role, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Guest:
      type: object
      required: [id, name, phone, nationalId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        email:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ReservationDetail:
      type: object
      required: [id, roomId, guestId, employeeId, checkInDate, checkOutDate, status, totalAmount, paymentReceiptNumber, createdAt, room, guest, employee]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        guestId:
          type: integer
        employeeId:
          type: integer
        checkInDate:
          type: string
        checkOutDate:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        paymentReceiptNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        room:
          $ref: "#/components/schemas/Room"
        guest:
          $ref: "#/components/schemas/Guest"
        employee:
          $ref: "#/components/schemas/Employee"
    ReservationInput:
      type: object
      required: [roomId, guestId, employeeId, checkInDate, checkOutDate, paymentReceiptNumber]
      properties:
        roomId:
          type: integer
        guestId:
          type: integer
        employeeId:
          type: integer
        checkInDate:
          type: string
        checkOutDate:
          type: string
        totalAmount:
          type: number
        paymentReceiptNumber:
          type: string
        notes:
          type: string
    ReservationUpdate:
      type: object
      properties:
        roomId:
          type: integer
        guestId:
          type: integer
        employeeId:
          type: integer
        checkInDate:
          type: string
        checkOutDate:
          type: string
        totalAmount:
          type: number
        paymentReceiptNumber:
          type: string
        notes:
          type: string
        status:
          type: string
...