# lib/api-zod/src/generated/types/reservationDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';
import type { Guest } from './guest';
import type { Room } from './room';

export interface ReservationDetail {
  id: number;
  roomId: number;
  guestId: number;
  employeeId: number;
  checkInDate: string;
  checkOutDate: string;
  status: string;
  totalAmount: number;
  paymentReceiptNumber: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  room: Room;
  guest: Guest;
  employee: Employee;
}