# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  roomId: number;
  guestId: number;
  employeeId: number;
  checkInDate: string;
  checkOutDate: string;
  totalAmount?: number;
  paymentReceiptNumber: string;
  notes?: string;
}