# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  number: string;
  type: string;
  floor: number;
  status: string;
  pricePerNight: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}